remove bit_cast/bit_helper
