+++ be/src/vec/functions/function_variadic_arguments.h
@@ -20,7 +20,6 @@
 
 #include "vec/columns/column_string.h"
 #include "vec/columns/column_vector.h"
-#include "vec/common/bit_cast.h"
 #include "vec/data_types/data_type.h"
 #include "vec/data_types/data_type_decimal.h"
 #include "vec/data_types/data_type_number.h"
